:art: polished up the post gen script

+++ hooks/post_gen_project.sh
@@ -1,17 +1,29 @@
 #!/usr/bin/env bash
 
-if [[ "{{cookiecutter.open_source_license}}" = "No License" ]]
+if [[ "{{cookiecutter.docker_development}}" = "yes" ]]
 then
-    rm LICENSE
+    docker-compose up -d
+
+    echo
+    echo "Docker has composed the application and is executing at http://localhost:5000."
+else
+    if [[ "{{cookiecutter.open_source_license}}" = "No License" ]]
+    then
+        rm LICENSE
+    fi
+    echo "Preparing web application..."
+    python3 -m pip install pipenv
+
+    echo "Fetching dependencies"
+    pipenv install
+
+    echo
+    echo "Starting web application"
+    pipenv run python manage.py runserver
 fi
 
-echo "Preparing web application..."
-python3 -m pip install pipenv
 
-echo "Fetching dependencies"
-pipenv install
 
-echo
-echo "Docker has composed the application and is executing at http://localhost:5000."
+
